Fix Coverity 1503218: negative loop bound
OPENSSL_sk_num returns an integer which can theoretically be negative.
Assigning this to a size_t and using it as a loop bound isn't ideal.

Rather than adding checked for NULL or negative returns, changing the loop
index and end to int is simpler.

(Merged from https://github.com/openssl/openssl/pull/17954)

(cherry picked from commit 0bcae9893b99666158dd8b35fb674e6188b0b5c3)

+++ crypto/encode_decode/decoder_pkey.c
@@ -259,7 +259,7 @@ static void collect_decoder(OSSL_DECODER *decoder, void *arg)
 {
     struct collect_data_st *data = arg;
     STACK_OF(EVP_KEYMGMT) *keymgmts = data->keymgmts;
-    size_t i, end_i;
+    int i, end_i;
     EVP_KEYMGMT *keymgmt;
     const OSSL_PROVIDER *prov;
     void *provctx;
